added API XmlReport methods

+++ app/models/ReportInfo.scala
@@ -5,6 +5,6 @@ package models
 /* output List[T] */
 
 case class ReportInfo(
-  val ReportID: Int,
+  val ReportID: Long,
   val Url: Option[String] = None,
   val StatusReport: String)
